fix build warning

+++ engine/src/mysql/sql.cpp
@@ -220,7 +220,7 @@ int32_t Sql::query(const char *stmt, size_t size)
     /* Client error message numbers are listed in the MySQL errmsg.h header
      * file. Server error message numbers are listed in mysqld_error.h
      */
-    if (EXPECT_FALSE(mysql_real_query(_conn, stmt, size)))
+    if (EXPECT_FALSE(mysql_real_query(_conn, stmt, (unsigned long)size)))
     {
         int32_t e = mysql_errno(_conn);
         if (CR_SERVER_LOST != e && CR_SERVER_GONE_ERROR != e)
@@ -229,7 +229,8 @@ int32_t Sql::query(const char *stmt, size_t size)
         }
 
         // reconnect and try again
-        if (mysql_ping(_conn) || mysql_real_query(_conn, stmt, size))
+        if (mysql_ping(_conn)
+            || mysql_real_query(_conn, stmt, (unsigned long)size))
         {
             return mysql_errno(_conn);
         }
